Added screenshots table

+++ src/main/java/com/napier/sem/Country.java
@@ -1,3 +1,3 @@
 package com.napier.sem;
 
-public record Country(String code, String name, String continent, String region, int population, String capital) {}
+public record Country(String code, String name, String continent, String region, long population, String capital) {}
